scratchlivedb-tool: Add --require-base
Ensure that every entry in the db shares the same passed in base.

All my music is in C:/music/, which translates to a base of music/ in
the db. However if someone adds something from my desktop to the db,
the syncing logic doesn't find a shared base, and tries to sync everything
relative to /, which causes add/remove of all the entries.